refactor(telegram): extract ops/info slash commands from gateway (#2996)
Move /doctor /grant /dangerous /permissions /version /whoami /commands
(plus the formatWhoamiCard helper) verbatim into bot-commands-ops-info.ts
behind registerOpsInfoCommands(bot, deps). Read-only info surfaces, no
send-path or turn-lifecycle coupling; every gateway-local helper is
injected via deps (factory-deps precedent), module reads no globals.
Registration stays at the original call site to preserve grammy's
in-order command matching. Phase 5 leaf move.
